+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [0.19.0] - 2021-05-31
+
+- Added connection options `access_token` and `refresh_token` to support OpenID Connect in [WebSocket Protocol V3](https://github.com/exasol/websocket-api/blob/master/docs/WebsocketAPIV3.md).
+- PyEXASOL default protocol version will be upgraded to `3` if connection option `access_token` or `refresh_token` were used.
+
 ## [0.18.1] - 2021-03-15
 
 - Fixed orphan process check in HTTP Transport being enabled on Windows instead of POSIX OS.